[AdminCampingRegisterPage] 컴포넌트로 만들어서 폴더에 뺴둠
